Adds javadocs for methods and classes. Changes port for service from 8080 to 26501.
